upgrade to typescript; change to babel 7; add prettier;
