+++ .prettierrc
@@ -1,0 +1,5 @@
+{
+    "tabWidth": 4,
+    "useTabs": false,
+    "trailingComma": "all"
+}
